This is the first attempt  at getting rid of all the size_t errors that show up when building using the Windows 7 SDK.  A special thanks to Arnd Menge for pointing this out and including a Makefile and instructions for the SDK that exposed them all.  For whatever reason, the build process using the DDK does not show all the errors.
I did not attempt to remove "int" entirely from the program yet, but only the ones where it is compared and converted to other variables.  More work will be required to completely get rid of them, but I didn't want to completely break everything.

As it is, it exposed at least two bad (core dump) issues where we were working on data passed without checking if length>0, so this has been a very good exercise so far.  Thanks again Arnd.          

Of course thanks also to Nate Lawson who contributed the new USB support- I hope to convert soon and finally get rid of this parallel port, since I lose a few seconds on my Windows clock each time a disk is read or written. :)

Of course thanks to Spiro and Womo as always for continuing to develop OpenCBM!


Pete Rittwage




git-svn-id: https://c64preservation.com/svn/nibtools/trunk@248 048eaad1-c872-4173-8c30-523bb29f53ec

+++ prot.c
@@ -47,7 +47,7 @@ void shift_buffer_right(BYTE *buffer, int length, int n)
 }
 
 BYTE *
-align_vmax(BYTE * work_buffer, int tracklen)
+align_vmax(BYTE * work_buffer, size_t tracklen)
 {
 	BYTE *pos, *buffer_end, *start_pos;
 	int run;
@@ -77,7 +77,7 @@ align_vmax(BYTE * work_buffer, int tracklen)
 }
 
 BYTE *
-align_vmax_new(BYTE * work_buffer, int tracklen)
+align_vmax_new(BYTE * work_buffer, size_t tracklen)
 {
 	BYTE *pos, *buffer_end, *key_temp, *key;
 	int run, longest;
@@ -113,7 +113,7 @@ align_vmax_new(BYTE * work_buffer, int tracklen)
 }
 
 BYTE *
-align_vmax_cw(BYTE * work_buffer, int tracklen)
+align_vmax_cw(BYTE * work_buffer, size_t tracklen)
 {
 	BYTE *pos, *buffer_end, *start_pos;
 	int run;
@@ -140,7 +140,7 @@ align_vmax_cw(BYTE * work_buffer, int tracklen)
 // Line up the track cycle to the start of the longest gap mark
 // this helps some custom protection tracks master properly
 BYTE *
-auto_gap(BYTE * work_buffer, int tracklen)
+auto_gap(BYTE * work_buffer, size_t tracklen)
 {
 	BYTE *pos, *buffer_end, *key_temp, *key;
 	int run, longest;
@@ -185,7 +185,7 @@ auto_gap(BYTE * work_buffer, int tracklen)
 // we can line up the track cycle to this
 // in lieu of no other hints
 BYTE *
-find_bad_gap(BYTE * work_buffer, int tracklen)
+find_bad_gap(BYTE * work_buffer, size_t tracklen)
 {
 	BYTE *pos, *buffer_end, *key_temp, *key;
 	int run, longest;
@@ -225,7 +225,7 @@ find_bad_gap(BYTE * work_buffer, int tracklen)
 // Line up the track cycle to the start of the longest sync mark
 // this helps some custom protection tracks master properly
 BYTE *
-find_long_sync(BYTE * work_buffer, int tracklen)
+find_long_sync(BYTE * work_buffer, size_t tracklen)
 {
 	BYTE *pos, *buffer_end, *key_temp, *key;
 	int run, longest;
@@ -264,7 +264,7 @@ find_long_sync(BYTE * work_buffer, int tracklen)
 
 #include <assert.h>
 
-void fix_first_gcr(BYTE *gcrdata, int length, int pos)
+void fix_first_gcr(BYTE *gcrdata, size_t length, size_t pos)
 {
     // fix first bad byte in a row
     unsigned int lastbyte, mask, data;
@@ -286,7 +286,7 @@ void fix_first_gcr(BYTE *gcrdata, int length, int pos)
 }
 
 
-void fix_last_gcr(BYTE *gcrdata, int length, int pos)
+void fix_last_gcr(BYTE *gcrdata, size_t length, size_t pos)
 {
     // fix last bad byte in a row
     unsigned int lastbyte, mask, data;
